+++ Cython/Shadow.py
@@ -121,10 +121,13 @@ overflowcheck.fold = optimization.use_switch = \
 final = internal = type_version_tag = no_gc_clear = no_gc = _empty_decorator
 
 
+_cython_inline = None
 def inline(f, *args, **kwds):
     if isinstance(f, basestring):
-        from Cython.Build.Inline import cython_inline
-        return cython_inline(f, *args, **kwds)
+        global _cython_inline
+        if _cython_inline is None:
+            from Cython.Build.Inline import cython_inline as _cython_inline
+        return _cython_inline(f, *args, **kwds)
     else:
         assert len(args) == len(kwds) == 0
         return f